json mode works, in detail
- user-api-token merged @ub cheers
- cleaned up
- moved autoedit to api_smp w/ separate register and url_prefix in __init__
- updated README with X-Auth header and json calls for curl

open issues:
 - list files with files/ (think need GET/POST switch in the '/files' route
 - download files with files/data/result.wav doesnt work even though path exists. weird 404 logic with that
 - api.before_request global? double definition in api and api_smp
 - __user doesnt work, python foo, changed to _user
