+++ README.md
@@ -46,25 +46,16 @@ API
 
 uploading file with curl
 
-`curl -v -F "soundfile=@\"./trk010-3.mp3\"" -F "dt_session=zniz"  http://127.0.0.1:5000/upload`
+`curl -H "X-Authentication: yuOJX-8paOqRJR8iefr7vL-Ozu5owbSUtr8SIM0K1L1EKPB9mWjPM52nydMEFyl7" -F "soundfile=@\"./mp3/trk009-3.mp3\"" http://127.0.0.1:5000/files`
 
-calling item with curl
+downloading file with curl
 
-`curl -v -X POST --data "dt_item=zniz_ul_trk011-3.mp3" http://127.0.0.1:5000/item`
+`curl -H "X-Authentication: yuOJX-8paOqRJR8iefr7vL-Ozu5owbSUtr8SIM0K1L1EKPB9mWjPM52nydMEFyl7" http://127.0.0.1:5000/files/trk013-4.mp3`
 
-calling download with dt_item
+calling autoedit with single file input with
 
-`curl -v -F "dt_item=zniz/uploaded/data/trk010-3-autoedit-5.wav" http://127.0.0.1:5000/download --output trk010-3-autoedit-5.wav`
+`curl -H "X-Authentication: yuOJX-8paOqRJR8iefr7vL-Ozu5owbSUtr8SIM0K1L1EKPB9mWjPM52nydMEFyl7" -H "Content-Type: application/json" -X POST --data '{"filenames": ["trk013-4.mp3"]}' http://127.0.0.1:5000/api/smp/autoedit`
 
-calling autoedit from the command line with single file input
+downloading result file with curl
 
-`curl -v -X POST --data "dt_item=zniz/uploaded/trk010-3.mp3" http://127.0.0.1:5000/autoedit`
-
-calling autoedit from the command line with multiple file inputs
-
-`curl -v -F "dt_item[]=zniz/uploaded/trk010-3.mp3" -F "dt_item[]=zniz/uploaded/trk011-3.mp3" http://127.0.0.1:5000/autoedit`
-
-UNTESTED: calling autoedit with browser GET
-
-<http://127.0.0.1:5000/autoedit?dt_item=zniz_ul_trk011-3.mp3>
-
+`curl -H "X-Authentication: yuOJX-8paOqRJR8iefr7vL-Ozu5owbSUtr8SIM0K1L1EKPB9mWjPM52nydMEFyl7" http://127.0.0.1:5000/files/data/trk013-4-autoedit-0.wav`
